third_party/blink/web_tests/platform/generic/external/wpt/css/css-backgrounds/parsing/background-position-computed-expected.txt

This is a testharness.js-based test.
PASS Property background-position value '1px'
PASS Property background-position value '1px center'
PASS Property background-position value '-2% -3%'
PASS Property background-position value '5% top'
PASS Property background-position value 'center'
PASS Property background-position value 'center center'
PASS Property background-position value 'center 6px'
PASS Property background-position value 'center left'
FAIL Property background-position value 'center right 7%' assert_equals: expected "93% 50%" but got "right 7% 50%"
PASS Property background-position value 'center bottom'
FAIL Property background-position value 'center top 8px' assert_equals: expected "50% 8px" but got "50% top 8px"
PASS Property background-position value 'left'
PASS Property background-position value 'right 9%'
FAIL Property background-position value 'left 10px center' assert_equals: expected "10px 50%" but got "left 10px 50%"
FAIL Property background-position value 'right 11% bottom' assert_equals: expected "89% 100%" but got "right 11% 100%"
FAIL Property background-position value 'left 12px top 13px' assert_equals: expected "12px 13px" but got "left 12px top 13px"
PASS Property background-position value 'right center'
PASS Property background-position value 'left bottom'
FAIL Property background-position value 'right top 14%' assert_equals: expected "100% 14%" but got "100% top 14%"
PASS Property background-position value 'bottom'
FAIL Property background-position value 'top 15px center' assert_equals: expected "50% 15px" but got "50% top 15px"
FAIL Property background-position value 'bottom 16% left' assert_equals: expected "0% 84%" but got "0% bottom 16%"
FAIL Property background-position value 'top 17px right -18px' assert_equals: expected "calc(100% + 18px) 17px" but got "right -18px top 17px"
PASS Property background-position value 'bottom center'
PASS Property background-position value 'top left'
FAIL Property background-position value 'bottom right 19%' assert_equals: expected "81% 100%" but got "right 19% 100%"
PASS Property background-position value 'calc(10px + 0.5em) calc(10px - 0.5em)'
PASS Property background-position value 'calc(10px - 0.5em) calc(10px + 0.5em)'
PASS Property background-position value '12px 13px, 50% 6px, 30px -10px'
Harness: the test ran to completion.
